refactoring

+++ src/main/java/technikal/task/fishmarket/services/FishRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import technikal.task.fishmarket.models.Fish;
 
-public interface FishRepository extends JpaRepository<Fish, Integer> {
+public interface FishRepository extends JpaRepository<Fish, Long> {
 
 }
